Some provisioning fixes and modified to use Amazon VPC s we can configure /etc/hosts on the nodes automatically (using vagrant-hostmanager)

+++ sync/scripts/jumphost.sh
@@ -2,12 +2,19 @@
 
 set -e
 
+test -f /tmp/$(basename $0).run && exit 0
+
+PE_VER=3.2.3
+PE_TAR=puppet-enterprise-${PE_VER}-el-6-x86_64.tar.gz
+
+curl -sLo /vagrant/pe-install/${PE_TAR} https://s3.amazonaws.com/pe-builds/released/${PE_VER}/${PE_TAR}
+
 cp /vagrant/keys/jump_key /root/.ssh/id_rsa
 
 yum -y -q install gcc wget python python-devel git gpm-devel
 easy_install fabric
 
-curl https://raw.githubusercontent.com/fesplugas/rbenv-installer/master/bin/rbenv-installer | bash
+curl -s https://raw.githubusercontent.com/fesplugas/rbenv-installer/master/bin/rbenv-installer | bash
 
 cat >> /root/.bashrc <<"EOF"
 
@@ -24,7 +31,8 @@ EOF
 
 source /root/.bashrc
 
-rbenv install 1.9.3-p547
+rbenv install --skip-existing 1.9.3-p547
 rbenv global 1.9.3-p547
 
+touch /tmp/$(basename $0).run
